requirements.txt
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
colorama==0.4.4
cryptography==36.0.1
docutils==0.18.1
filetype==1.0.10
idna==3.3
importlib-metadata==4.10.1
jeepney==0.7.1
keyring==23.5.0
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.1
pkginfo==1.8.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
readme-renderer==32.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
SecretStorage==3.3.1
six==1.16.0
tqdm==4.62.3
twine==3.8.0
urllib3==1.26.8
webencodings==0.5.1
zipp==3.7.0
